19_iot_anomaly_detection/requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
pyarrow>=14.0
pyyaml>=6.0
matplotlib>=3.7
pytest>=7.0
